+++ pawpals/src/main/java/com/pawpals/libs/DBType.java
@@ -1,4 +1,4 @@
-package com.pawpals.interfaces;
+package com.pawpals.libs;
 
 public enum DBType {
 	MYSQL, HSQLDB
